windows error

+++ tests/test_models/test_archs/test_wrapper.py
@@ -1,8 +1,8 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 import os
 import os.path as osp
+import platform
 import shutil
-import stat
 from unittest import TestCase
 from unittest.mock import MagicMock
 
@@ -73,12 +73,9 @@ class TestWrapper(TestCase):
         model.init_weights()
 
         # delete saved model to save space
-        if not os.access(model_path, os.W_OK):
-            os.chmod(model_path, stat.S_IWUSR)
-        shutil.rmtree(model_path)
-        if not os.access(ckpt_path, os.W_OK):
-            os.chmod(ckpt_path, stat.S_IWUSR)
-        shutil.rmtree(ckpt_path)
+        if 'win' not in platform.system().lower():
+            shutil.rmtree(model_path)
+            shutil.rmtree(ckpt_path)
 
         # 4. test loading without repo_id
         model = MODELS.build(
